Modèle de base des fiches (DB + Backend) Service de gestion des fiches (Backend) WS de gestion loggué des fiches (Backend) WS de visionnage publique des fiches (Backend)

+++ src/main/java/be/arbiter/clouddragonsheet/configuration/security/services/UserDetailsImpl.java
@@ -21,7 +21,7 @@ public class UserDetailsImpl implements UserDetails {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
 
     private String username;
 
@@ -34,7 +34,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String username, String email, String password,boolean enabled,
+    public UserDetailsImpl(Integer id, String username, String email, String password,boolean enabled,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
